requirements.txt
PyQt6
PyQt6-WebEngine
openpyxl
google-generativeai
mysql-connector-python
Pillow